+++ yospace/src/main/java/com/bitmovin/player/integration/yospace/OnTruexAdFreeListener.kt
@@ -1,7 +1,0 @@
-package com.bitmovin.player.integration.yospace
-
-import com.bitmovin.player.api.event.listener.EventListener
-
-interface OnTruexAdFreeListener : EventListener<TruexAdFreeEvent> {
-    fun onTruexAdFree(event: TruexAdFreeEvent)
-}