+++ yospace/src/main/java/com/bitmovin/player/integration/yospace/TruexAdFreeEvent.kt
@@ -1,5 +1,3 @@
 package com.bitmovin.player.integration.yospace
 
-import com.bitmovin.player.api.event.data.BitmovinPlayerEvent
-
-class TruexAdFreeEvent : BitmovinPlayerEvent()
+class TruexAdFreeEvent : CustomEvent()