+++ yospace/src/main/java/com/bitmovin/player/integration/yospace/config/YospaceConfig.kt
@@ -2,11 +2,16 @@ package com.bitmovin.player.integration.yospace.config
 
 import com.bitmovin.player.integration.yospace.YospaceLiveInitialisationType
 
-data class YospaceConfiguration(
+data class YospaceConfig(
     val userAgent: String? = null,
     val readTimeout: Int = 25_000,
     val connectTimeout: Int = 25_000,
     val requestTimeout: Int = 25_000,
     val liveInitialisationType: YospaceLiveInitialisationType = YospaceLiveInitialisationType.DIRECT,
-    val debug: Boolean = false
+    val isDebug: Boolean = false,
+    val filterMetadataType: MetadataType? = MetadataType.EMSG
 )
+
+public enum class MetadataType{
+    ID3, EMSG
+}