+++ yospace/src/main/java/com/bitmovin/player/integration/yospace/config/YospaceSourceConfig.kt
@@ -2,7 +2,7 @@ package com.bitmovin.player.integration.yospace.config
 
 import com.bitmovin.player.integration.yospace.YospaceAssetType
 
-data class YospaceSourceConfiguration(
+data class YospaceSourceConfig(
     val assetType: YospaceAssetType,
     val retryExcludingYospace: Boolean = false
 )