+++ yospacesample/src/main/java/com/bitmovin/player/integration/yospacesample/Stream.kt
@@ -1,12 +1,12 @@
 package com.bitmovin.player.integration.yospacesample
 
-import com.bitmovin.player.integration.yospace.config.TruexConfiguration
-import com.bitmovin.player.integration.yospace.config.YospaceSourceConfiguration
+import com.bitmovin.player.integration.yospace.config.TruexConfig
+import com.bitmovin.player.integration.yospace.config.YospaceSourceConfig
 
 data class Stream(
     val title: String,
     val contentUrl: String,
     val drmUrl: String? = null,
-    val yospaceSourceConfig: YospaceSourceConfiguration,
-    val truexConfig: TruexConfiguration? = null
+    val yospaceSourceConfig: YospaceSourceConfig,
+    val truexConfig: TruexConfig? = null
 )